refactor: split LimitReader into limitreader.py
