+++ pywb/webagg/inputrequest.py
@@ -1,5 +1,5 @@
 from pywb.utils.loaders import extract_post_query, append_post_query
-from pywb.utils.loaders import LimitReader
+from pywb.utils.limitreader import LimitReader
 from pywb.utils.statusandheaders import StatusAndHeadersParser
 
 from six.moves.urllib.parse import urlsplit, quote
